backend/requirements.txt
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==4.3.0
certifi==2025.11.12
click==8.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.124.4
fastapi-cli==0.0.16
fastapi-cloud-cli==0.6.0
fastar==0.8.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
orjson==3.11.5
passlib==1.7.4
psycopg2==2.9.11
pyasn1==0.6.1
pydantic==2.12.5
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
rsa==4.9.1
sentry-sdk==2.47.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.6.2
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
